Statistics: Number of wins saving correctly
Just now, race conditions started appearing during the testing process. The lobby/non-host lobby seem to be competing for loading on the client side. This just started so I don't believe it has anything to do with my added functionality, as it is basically the same as what I've added in my prior two commits.